rename colors to css_variables
