+++ src/renderer/components/labelled_text.tsx
@@ -1,7 +1,7 @@
 import { css } from "@emotion/react";
 import React from "react";
 
-import { cssVar } from "@/styles/colors";
+import { cssVar } from "@/styles/css_variables";
 import { withFont } from "@/styles/with_font";
 
 type LabelledTextProps = {
