+++ src/renderer/pages/settings/settings_page.tsx
@@ -17,7 +17,7 @@ import { DualPane } from "@/components/dual_pane";
 import { useMousetrap } from "@/lib/hooks/use_mousetrap";
 import { useSettingsModal } from "@/lib/hooks/use_settings_modal";
 import type { SettingSection } from "@/pages/settings/types";
-import { cssVar } from "@/styles/colors";
+import { cssVar } from "@/styles/css_variables";
 import { platformTitleBarStyles } from "@/styles/platform_title_bar_styles";
 import { withSlippiBackground } from "@/styles/with_slippi_background";
