+++ src/renderer/pages/spectate/share_gameplay_block/broadcast_panel.tsx
@@ -11,7 +11,7 @@ import { makeIntlFormatter } from "react-timeago/defaultFormatter";
 import { useAppStore } from "@/lib/hooks/use_app_store";
 import { getLocale } from "@/lib/time";
 import type { SupportedLanguage } from "@/services/i18n/util";
-import { cssVar } from "@/styles/colors";
+import { cssVar } from "@/styles/css_variables";
 
 import { BroadcastPanelMessages as Messages } from "./broadcast_panel.messages";
 import { StartBroadcastDialog } from "./start_broadcast_dialog";
