+++ src/renderer/pages/spectate/spectator_id_block/spectator_id_block.tsx
@@ -4,7 +4,7 @@ import InputBase from "@mui/material/InputBase";
 import React from "react";
 
 import { InfoBlock } from "@/components/info_block";
-import { cssVar } from "@/styles/colors";
+import { cssVar } from "@/styles/css_variables";
 
 import { SpectatorIdBlockMessages as Messages } from "./spectator_id_block.messages";
